Update patch set 1
Patch Set 1: -Code-Review

I never really looked into the 'why' for the overlay, just noticed this element was missing...

It seems previously instead of 'bootloader', 'fastboot' was listed.  But now, it seems there is actually no need for this overlay.

https://github.com/CyanogenMod/android_device_motorola_qcom-common/commit/40d58ec1ea4c328b466bcaf907d1406f4ac42f09

I'll adjust the patch to remove the overlay entirely.

Patch-set: 1