Update patch set 1
Patch Set 1:

Yes, audio effects can't work when tunnel is used. Combined with non-consistent tunnel creation behavior - sometimes tunnel failed on track change and playback fell back to LPA for a track, it created not very good user experience:
When user enabled e.g. EQ, it didn't work for most tracks in a playlist, but the EQ occasionally kicked in for a random track (when LPA fall back happen). There was a patch trying to fix the tunnel bugs in frameworks_av, but if I remember correctly, it was never quite finished.

Patch-set: 1
Label: Code-Review=0