Update patch set 1
Patch Set 1: Code-Review+1

Emerson: the only motorola/qcom-common users left are Motorola msm8960 2012 models.

On non US model, the utility will report that US GSM is already unlocked and there's no need to change anything (tested on my Euro xt925).

The primary targets for this utility are xt897 (jbbl), xt907 and xt926 (kkbl).

There's no intention to run it automatically. It's just meant to be there, ready for advanced users to run it from a terminal app or adb shell, if they decide to.

Patch-set: 1