Update patch set 1
Patch Set 1:

Dave: This is about the behavior when exiting the screensaver/dream state (when dreaming is enabled while the device is connected to power or docked) by tap on the screen.
The current behavior is that tapping on the screen when dreaming ends the screensaver and the screen goes off and the device locks, while the expected behavior would be to wake from the dream and turn the display on by the tap.

Nevertheless, this change is currently wrong as it leads to display never turning off after user inactivity timeout (regardless whether dream is enabled or not, powered/docked or not...)

Patch-set: 1