Update patch set 1
Patch Set 1:

While this was working fine during the last autumn for me, later I started to have issues with microphone not being heard by the remote party during VOIP calls. So it's VOIP what needs some love. I haven't  found the time to look into it yet, but I hope to be able to deal with it soon.

Patch-set: 1
Label: Verified=0